Attempts to use a different conversion library

+++ ci/postinstall/install-resources-data.ts
@@ -1,6 +1,7 @@
 import * as fs from 'fs';
 import StreamZip from 'node-stream-zip';
-import psd2png from 'psd2png';
+import { readPsd } from 'ag-psd';
+import 'ag-psd/initialize-canvas';
 import { Agent } from 'undici';
 import { getSeperatedValueData, recreateDirectory } from '../utils';
 import {
@@ -36,7 +37,7 @@ const downloadFile = async (url: string, dir: string) => {
 recreateDirectory(INSTALLED_RESOURCES_DIR);
 
 // Bootswatch theme for Bootstrap UI: https://bootswatch.com/cerulean/
-await downloadFile("https://bootswatch.com/5/cerulean/bootstrap.min.css", INSTALLED_RESOURCES_DIR);
+/*await downloadFile("https://bootswatch.com/5/cerulean/bootstrap.min.css", INSTALLED_RESOURCES_DIR);
 
 // Bootstrap UI components & framework: https://getbootstrap.com/docs/5.0/getting-started/introduction/
 // Minified bundle downloaded from CDN: https://getbootstrap.com/docs/5.3/getting-started/download/#cdn-via-jsdelivr
@@ -122,7 +123,7 @@ const signWritingAlphabets = await Promise.all(
     })
 );
 fs.writeFileSync(SIGN_WRITING_ALPHABETS_FILE_PATH, JSON.stringify(signWritingAlphabets));
-console.log(`=> Wrote ${signWritingDictionaries.length} SignWriting alphabets to: ${SIGN_WRITING_ALPHABETS_FILE_PATH}`);
+console.log(`=> Wrote ${signWritingDictionaries.length} SignWriting alphabets to: ${SIGN_WRITING_ALPHABETS_FILE_PATH}`);*/
 
 // Hand pictures
 const HAND_PICTURES_INSTALL_DIR = `${INSTALLED_RESOURCES_DIR}/${HAND_PICTURES_DIR}`;
@@ -145,8 +146,16 @@ const getHandPictures = (baseSymbolId: string, orientationNumber: number): Promi
 const convertPsdToPng = (dirPath: string, fileName: string) => {
     const psdFilePath = `${dirPath}/${fileName}.psd`;
     const buffer = fs.readFileSync(psdFilePath);
-    const pngBuffer = psd2png(buffer);
-    fs.writeFileSync(`${dirPath}/${fileName}.png`, pngBuffer);
+    const psd = readPsd(buffer, {skipThumbnail: true});
+    if (psd && psd.children) {
+        console.log(psd.children[0].canvas);
+    } else {
+        console.log("undefined!!!");
+    }
+    if (!psd || !psd.children) {
+        throw new Error(`Failed to parse ${psdFilePath}`);
+    }
+    fs.writeFileSync(`${dirPath}/${fileName}.png`, (psd.children[0].canvas as any).toBuffer());
     fs.rmSync(psdFilePath);
 };
